+++ cmd/issue-cli/cmd_done.go
@@ -25,7 +25,7 @@ func runDone(ctx *Context, args []string) error {
 		return err
 	}
 
-	issue, _, err := findIssueOrErr(ctx.Project, slug)
+	issue, _, err := findIssueOrErr(ctx, slug)
 	if err != nil {
 		return err
 	}