+++ cmd/issue-cli/cmd_unclaim.go
@@ -27,7 +27,7 @@ func runUnclaim(ctx *Context, args []string) error {
 		return err
 	}
 
-	issue, _, err := findIssueOrErr(ctx.Project, slug)
+	issue, _, err := findIssueOrErr(ctx, slug)
 	if err != nil {
 		return err
 	}